// starfall-viewer/src/types/PointSourceDetails.ts
/*************************************************************************************************
* Licensed to the Apache Software Foundation (ASF) under one
* or more contributor license agreements.  See the NOTICE file
* distributed with this work for additional information
* regarding copyright ownership.  The ASF licenses this file
* to you under the Apache License, Version 2.0 (the
* "License"); you may not use this file except in compliance
* with the License.  You may obtain a copy of the License at
*
*     http://www.apache.org/licenses/LICENSE-2.0
*
* Unless required by applicable law or agreed to in writing,
* software distributed under the License is distributed on an
* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
* KIND, either express or implied.  See the License for the
* specific language governing permissions and limitations
* under the License.
**************************************************************************************************/

export type PointSourceDetails = {
  point_source_id: string;
  sighting_id: string;
  platform_name: string;
  sensor_name: string;
  time: number;
  intensity: number;
  cluster_size: number | null;
  above_horizon: boolean;
  meas_near_point_ecef_m: number[];
  meas_far_point_ecef_m: number[];
  tags: string[];
  [s: string]: unknown;
};